# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: Get all tasks
      tags: [tasks]
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      summary: Create a new task
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: The created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      summary: Patch task
      operationId: patchTaskById
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found

    delete:
      summary: Delete task
      operationId: deleteTaskById
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        "204":
          description: Deleted

  /users:
    get:
      summary: Get all users
      tags: [ users ]
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      summary: Create a new user
      tags: [ users ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: The created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    patch:
      summary: Patch user by id
      operationId: PatchUserById
      tags: [ users ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      summary: Delete user by id
      operationId: DeleteUserById
      tags: [ users ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        "204":
          description: Deleted


components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint64
        task:
          type: string
        is_done:
          type: boolean
      required: [id, task, is_done]

    TaskCreate:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: boolean
      required: [task, is_done]

    TaskPatch:
      type: object
      properties:
        task:
          type: string
          nullable: true
        is_done:
          type: boolean
          nullable: true

    User:
      type: object
      properties:
        id:
          type: integer
          format: uint64
        email:
          type: string
        password:
          type: string
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [ id, email, password, created_at, updated_at ]

    UserCreate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [ email, password ]

    UserPatch:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true